+++ changelog.md
@@ -1,5 +1,8 @@
 # UnderScript Changelog
 
+## Version 0.57.1 (2024-04-05)
+1. Fixed collection dropdown filter translation error
+
 ## Version 0.57.0 (2024-04-01)
 ### New Features
 1. Added filter for tribes